refactor(scan): centralize session tag policy
