+++ packages/cli/src/scan-refresh-worker-integration.test.ts
@@ -93,6 +93,8 @@ vi.mock("@codesesh/core/runtime", async (importOriginal) => {
     createRegisteredAgents: mocks.createRegisteredAgents,
     synchronizePricingGeneration: mocks.synchronizePricingGeneration,
     ensureSessionTagsSync: mocks.ensureSessionTagsSync,
+    hasStaleSessionTags: actual.hasStaleSessionTags,
+    inheritSessionTags: actual.inheritSessionTags,
     FileSystemSessionSource: mocks.FileSystemSessionSource,
     PRICING_CAPTURE_EPOCH: actual.PRICING_CAPTURE_EPOCH,
     buildAgentCacheMeta: actual.buildAgentCacheMeta,
@@ -100,7 +102,6 @@ vi.mock("@codesesh/core/runtime", async (importOriginal) => {
     computeSessionDiff: actual.computeSessionDiff,
     planAgentScan: actual.planAgentScan,
     sessionSignature: actual.sessionSignature,
-    SMART_TAG_CLASSIFIER_REVISION: "smart-tags-v1",
     sortSessions: actual.sortSessions,
     synchronizeSessionSources: actual.synchronizeSessionSources,
     setCoreDiagnostics: actual.setCoreDiagnostics,
