+++ packages/core/src/discovery/scanner.ts
@@ -1,11 +1,4 @@
-import { availableParallelism } from "node:os";
-import { Worker } from "node:worker_threads";
-import type {
-  IdentifiedSessionHead,
-  SessionDetail,
-  SessionHead,
-  SmartTag,
-} from "../types/index.js";
+import type { IdentifiedSessionHead, SessionHead } from "../types/index.js";
 import { assertIdentifiedSessionHead } from "../contract/session.js";
 import { assertSessionIdentity } from "../contract/session-reference.js";
 import { filterSessionTreeByActivityWindow } from "../contract/session-tree.js";
@@ -14,7 +7,6 @@ import type {
   AgentScanOptions,
   BaseAgent,
   EnumeratedSessionSourceCapability,
-  SessionCacheMeta,
   SessionSourceFailure,
 } from "../agents/index.js";
 import {
@@ -24,16 +16,8 @@ import {
   SessionScanError,
 } from "../agents/index.js";
 import { filterSessionsByProjectScope } from "../projects/index.js";
-import {
-  classifySessionTags,
-  getSmartTagSourceTimestamp,
-  isWorkerLogMessage,
-  perf,
-  SMART_TAG_CLASSIFIER_REVISION,
-  type WorkerLogMessage,
-} from "../utils/index.js";
+import { perf } from "../utils/index.js";
 import { getCoreDiagnostics } from "../utils/diagnostics.js";
-import { getPricingGeneration } from "../pricing/index.js";
 import {
   markAgentCacheInitialized,
   markAgentFullSyncCompleted,
@@ -48,6 +32,7 @@ import {
   buildSessionPersistenceDiff,
 } from "./orchestrate.js";
 import { inspectAgentRefresh, planAgentScan } from "./agent-scan-plan.js";
+import { ensureSessionTags } from "./session-tags.js";
 
 export interface ScanOptions {
   /** Filter to specific agent name(s) */
@@ -60,8 +45,6 @@ export interface ScanOptions {
   to?: number;
   /** Use cached scan results if available */
   useCache?: boolean;
-  /** Enable smart refresh (fast cache + background incremental scan) */
-  smartRefresh?: boolean;
   /** Return cached session heads without validating the filesystem */
   cacheOnly?: boolean;
   /** Persist scan results to the SQLite cache */
@@ -115,17 +98,6 @@ export interface AgentScanTiming {
   identity?: number;
   tags?: number;
   total: number;
-}
-
-export interface SessionTagTiming {
-  sessions: number;
-  cacheHits: number;
-  staleSessions: number;
-  failedSessions: number;
-  getSessionDataCalls: number;
-  getSessionDataMs: number;
-  classifySessionTagsCalls: number;
-  classifySessionTagsMs: number;
 }
 
 function buildAgentScanOptions(
@@ -235,13 +207,6 @@ interface FinalizeAgentScanContext {
   onProgress?: (progress: ScanProgress) => void;
 }
 
-interface SmartTagWorkerResult {
-  id: string;
-  tags?: SmartTag[];
-  sourceUpdatedAt?: number;
-  error?: string;
-}
-
 function saveCachedSessionDiff(
   agent: BaseAgent,
   cachedSessions: SessionHead[],
@@ -269,216 +234,6 @@ function saveCachedSessionDiff(
     });
   }
   return persisted;
-}
-
-function getSmartTagWorkerCount(sessionCount: number): number {
-  if (sessionCount < 50) return 1;
-  return Math.min(sessionCount, Math.max(1, Math.min(4, availableParallelism() - 1)));
-}
-
-function chunkSessions<T>(items: T[], chunkCount: number): T[][] {
-  const chunks = Array.from({ length: chunkCount }, () => [] as T[]);
-  items.forEach((item, index) => {
-    chunks[index % chunkCount]!.push(item);
-  });
-  return chunks.filter((chunk) => chunk.length > 0);
-}
-
-export function ensureSessionTagsSync(
-  agent: BaseAgent,
-  sessions: SessionHead[],
-  onProgress?: (processed: number, total: number) => void,
-  classifierRevision = SMART_TAG_CLASSIFIER_REVISION,
-): { sessions: SessionHead[]; changed: boolean; timing: SessionTagTiming } {
-  let changed = false;
-  let processed = 0;
-  const total = sessions.length;
-  const timing: SessionTagTiming = {
-    sessions: total,
-    cacheHits: 0,
-    staleSessions: 0,
-    failedSessions: 0,
-    getSessionDataCalls: 0,
-    getSessionDataMs: 0,
-    classifySessionTagsCalls: 0,
-    classifySessionTagsMs: 0,
-  };
-
-  const tagged = sessions.map((session) => {
-    const sourceUpdatedAt = session.time_updated ?? session.time_created;
-    const currentTags = Array.isArray(session.smart_tags) ? session.smart_tags : null;
-    if (
-      currentTags &&
-      session.smart_tags_source_updated_at === sourceUpdatedAt &&
-      session.smart_tags_classifier_revision === classifierRevision
-    ) {
-      timing.cacheHits += 1;
-      processed += 1;
-      onProgress?.(processed, total);
-      return session;
-    }
-
-    timing.staleSessions += 1;
-    try {
-      timing.getSessionDataCalls += 1;
-      const getSessionDataStartedAt = performance.now();
-      let data: SessionDetail;
-      try {
-        data = agent.getSessionData(session.reference.sessionId);
-      } finally {
-        timing.getSessionDataMs += performance.now() - getSessionDataStartedAt;
-      }
-
-      timing.classifySessionTagsCalls += 1;
-      const classifySessionTagsStartedAt = performance.now();
-      let tags: SmartTag[];
-      try {
-        tags = classifySessionTags(data);
-      } finally {
-        timing.classifySessionTagsMs += performance.now() - classifySessionTagsStartedAt;
-      }
-
-      changed = true;
-      return {
-        ...session,
-        smart_tags: tags,
-        smart_tags_source_updated_at: getSmartTagSourceTimestamp(data),
-        smart_tags_classifier_revision: classifierRevision,
-      };
-    } catch {
-      timing.failedSessions += 1;
-      return session;
-    } finally {
-      processed += 1;
-      onProgress?.(processed, total);
-    }
-  });
-
-  return { sessions: tagged, changed, timing };
-}
-
-const SMART_TAG_WORKER_TIMEOUT_MS = 300_000;
-
-async function classifySessionTagsInWorker(
-  workerUrl: URL | string,
-  agentName: string,
-  sessionIds: string[],
-  meta: Record<string, SessionCacheMeta>,
-): Promise<SmartTagWorkerResult[]> {
-  const worker = new Worker(workerUrl, {
-    workerData: {
-      pricingGenerationId: getPricingGeneration().id,
-      agentName,
-      sessionIds,
-      meta,
-    },
-  });
-  let timer: NodeJS.Timeout | undefined;
-  try {
-    return await new Promise<SmartTagWorkerResult[]>((resolveWorker, rejectWorker) => {
-      timer = setTimeout(() => {
-        rejectWorker(
-          new Error(`Smart tag worker timed out after ${SMART_TAG_WORKER_TIMEOUT_MS}ms`),
-        );
-      }, SMART_TAG_WORKER_TIMEOUT_MS);
-      worker.on("message", (message: unknown) => {
-        if (isWorkerLogMessage(message)) {
-          relayWorkerLogMessage(message);
-          return;
-        }
-        const results = message as SmartTagWorkerResult[];
-        // An empty answer to a non-empty request means the worker could not
-        // do the job at all (e.g. unresolvable agent) — fail over to the
-        // synchronous path instead of silently shipping untagged sessions.
-        if (results.length === 0 && sessionIds.length > 0) {
-          rejectWorker(new Error("Smart tag worker returned no results"));
-          return;
-        }
-        resolveWorker(results);
-      });
-      worker.once("error", rejectWorker);
-      // Any exit before a result message — including a clean code 0 — must
-      // reject; the old `code !== 0` guard left the promise pending forever
-      // and hung server startup.
-      worker.once("exit", (code) => {
-        rejectWorker(new Error(`Smart tag worker exited with code ${code} before responding`));
-      });
-    });
-  } finally {
-    if (timer) clearTimeout(timer);
-    worker.terminate().catch(() => {});
-  }
-}
-
-function relayWorkerLogMessage(message: WorkerLogMessage): void {
-  const detail = {
-    ...message.data,
-    worker_ts: message.ts,
-    worker_pid: message.pid,
-    worker_thread_id: message.threadId,
-    worker_level: message.level,
-  };
-  if (message.level === "warn" || message.level === "error") {
-    getCoreDiagnostics()?.warn(message.event, detail);
-    return;
-  }
-  getCoreDiagnostics()?.info?.(message.event, detail);
-}
-
-async function ensureSessionTags(
-  agent: BaseAgent,
-  sessions: SessionHead[],
-  workerUrl?: URL | string,
-): Promise<{ sessions: SessionHead[]; changed: boolean }> {
-  const staleSessions = sessions.filter((session) => {
-    const sourceUpdatedAt = session.time_updated ?? session.time_created;
-    const currentTags = Array.isArray(session.smart_tags) ? session.smart_tags : null;
-    return (
-      !currentTags ||
-      session.smart_tags_source_updated_at !== sourceUpdatedAt ||
-      session.smart_tags_classifier_revision !== SMART_TAG_CLASSIFIER_REVISION
-    );
-  });
-
-  if (staleSessions.length === 0) {
-    return { sessions, changed: false };
-  }
-
-  const workerCount = workerUrl ? getSmartTagWorkerCount(staleSessions.length) : 1;
-  if (workerCount <= 1) {
-    return ensureSessionTagsSync(agent, sessions);
-  }
-
-  const meta = buildAgentCacheMeta(agent);
-  try {
-    const results = (
-      await Promise.all(
-        chunkSessions(
-          staleSessions.map((session) => session.reference.sessionId),
-          workerCount,
-        ).map((sessionIds) =>
-          classifySessionTagsInWorker(workerUrl!, agent.name, sessionIds, meta),
-        ),
-      )
-    ).flat();
-    const resultMap = new Map(results.filter((item) => item.tags).map((item) => [item.id, item]));
-
-    return {
-      changed: resultMap.size > 0,
-      sessions: sessions.map((session) => {
-        const result = resultMap.get(session.reference.sessionId);
-        if (!result?.tags || result.sourceUpdatedAt == null) return session;
-        return {
-          ...session,
-          smart_tags: result.tags,
-          smart_tags_source_updated_at: result.sourceUpdatedAt,
-          smart_tags_classifier_revision: SMART_TAG_CLASSIFIER_REVISION,
-        };
-      }),
-    };
-  } catch {
-    return ensureSessionTagsSync(agent, sessions);
-  }
 }
 
 export async function finalizeAgentScan(
