+++ packages/core/src/runtime/index.ts
@@ -48,6 +48,8 @@ export {
   commitDurableSessionPublication,
   computeSessionDiff,
   ensureSessionTagsSync,
+  hasStaleSessionTags,
+  inheritSessionTags,
   getAgentFullSyncCursor,
   getAnalyticsRevision,
   getSearchProjectDirectory,
